cluster-spark/requirements.txt
ipython
pandas
pyarrow
numpy
pyspark
python-dotenv
boto3